+++ packages/common/src/contracts/events.generated.ts
@@ -3,9 +3,13 @@
 
 import type { AuthEvents } from "@monark/auth/contracts"
 import type { FeatureFlagsEvents } from "@monark/feature-flags/contracts"
+import type { OrganizationsEvents } from "@monark/organizations/contracts"
+import type { RbacEvents } from "@monark/rbac/contracts"
 import type { UsersEvents } from "@monark/users/contracts"
 
 export type DomainEvent =
   | AuthEvents
   | FeatureFlagsEvents
+  | OrganizationsEvents
+  | RbacEvents
   | UsersEvents