+++ services/api/src/trpc/app-router.generated.ts
@@ -5,12 +5,16 @@ import { router } from "@monark/common/trpc"
 
 import { authRouter } from "@monark/auth/server"
 import { featureFlagsRouter } from "@monark/feature-flags/server"
+import { organizationsRouter } from "@monark/organizations/server"
+import { rbacRouter } from "@monark/rbac/server"
 import { usersRouter } from "@monark/users/server"
 
 export const appRouter = router({
-  auth        : authRouter,
-  featureFlags: featureFlagsRouter,
-  users       : usersRouter,
+  auth         : authRouter,
+  featureFlags : featureFlagsRouter,
+  organizations: organizationsRouter,
+  rbac         : rbacRouter,
+  users        : usersRouter,
 })
 
 export type AppRouter = typeof appRouter